Update the test
# Conflicts:
#	core/merge_test.go
